+++ src/font/u8g2.rs
@@ -55,17 +55,17 @@ impl FontMetrics for FontRenderer {
     }
 
     fn vertical_metrics(&self) -> font::VMetrics {
-        let ascent = self.get_ascent() as i32;
-        let descent = -(self.get_descent() as i32);
+        let ascent = self.get_ascent() as i16;
+        let descent = -(self.get_descent() as i16);
         font::VMetrics {
             ascent,
             descent,
-            line_spacing: self.get_default_line_height() as i32 - ascent + descent,
+            line_spacing: self.get_default_line_height() as i16 - ascent + descent,
         }
     }
 
-    fn advance(&self, character: char) -> u32 {
+    fn advance(&self, character: char) -> u16 {
         self.get_rendered_dimensions(character, EgPoint::zero(), VerticalPosition::Top)
-            .map_or(0, |d| d.advance.x as u32)
+            .map_or(0, |d| d.advance.x as u16)
     }
 }
