+++ src/render/chart/point.rs
@@ -10,7 +10,7 @@ pub struct PointRenderable<const N: usize> {
     /// The computed pixel coordinates of each data point.
     pub points: heapless::Vec<(i16, i16), N>,
     /// The point diameter in pixels.
-    pub point_size: u32,
+    pub point_size: u16,
     /// The bounding frame of the chart area.
     pub frame: Rectangle,
 }
@@ -41,7 +41,7 @@ impl<const N: usize> AnimatedJoin for PointRenderable<N> {
             self.points[i].1 =
                 i16::interpolate(source.points[i].1, self.points[i].1, domain.factor);
         }
-        self.point_size = u32::interpolate(source.point_size, self.point_size, domain.factor);
+        self.point_size = u16::interpolate(source.point_size, self.point_size, domain.factor);
         self.frame = Rectangle::interpolate(source.frame.clone(), self.frame.clone(), domain.factor);
     }
 }
@@ -49,11 +49,11 @@ impl<const N: usize> AnimatedJoin for PointRenderable<N> {
 impl<const N: usize, C: Copy> Render<C> for PointRenderable<N> {
     fn render(&self, render_target: &mut impl RenderTarget<ColorFormat = C>, style: &C) {
         let brush = SolidBrush::new(*style);
-        let half = (self.point_size / 2) as i32;
+        let half = (self.point_size / 2) as i16;
 
         for &(px, py) in &self.points {
             let rect = Rectangle::new(
-                Point::new(i32::from(px) - half, i32::from(py) - half),
+                Point::new(i16::from(px) - half, i16::from(py) - half),
                 Size::new(self.point_size, self.point_size),
             );
             render_target.fill(LinearTransform::default(), &brush, None, &rect);
